Separated basic components (#12)
