+++ src/components/Navbar.js
@@ -1,0 +1,15 @@
+import Logo from '../logo.png';
+
+const Navbar = () => {
+  return <nav className="navbar" role="navigation" aria-label="main navigation">
+    <div className="navbar-brand">
+      <a className="navbar-item" href="/#">
+        <img src={Logo}
+             alt="logo" width="150"
+             height="30" />
+      </a>
+    </div>
+  </nav>
+}
+
+export default Navbar;
